refactor(frontend): Merapihkan kode agar lebih mudah di baca dan di maintenance
