+++ frontend/package.json
@@ -15,6 +15,7 @@
   "dependencies": {
     "@tailwindcss/vite": "^4.0.1",
     "clsx": "^2.1.1",
+    "lucide-react": "^0.483.0",
     "motion": "^12.0.6",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
